chore: move validationRule to the data plane utils in core
