+++ extensions/data-plane/data-plane-gcs/src/main/java/org/eclipse/dataspaceconnector/gcp/dataplane/gcs/GcsDataSinkFactory.java
@@ -17,13 +17,13 @@ package org.eclipse.dataspaceconnector.gcp.dataplane.gcs;
 import com.google.auth.oauth2.AccessToken;
 import com.google.auth.oauth2.GoogleCredentials;
 import com.google.cloud.storage.StorageOptions;
+import org.eclipse.dataspaceconnector.dataplane.common.validation.ValidationRule;
 import org.eclipse.dataspaceconnector.dataplane.spi.pipeline.DataSink;
 import org.eclipse.dataspaceconnector.dataplane.spi.pipeline.DataSinkFactory;
 import org.eclipse.dataspaceconnector.gcp.core.common.GcpAccessToken;
 import org.eclipse.dataspaceconnector.gcp.core.common.GcpException;
 import org.eclipse.dataspaceconnector.gcp.core.storage.GcsStoreSchema;
 import org.eclipse.dataspaceconnector.gcp.dataplane.gcs.validation.GcsSinkDataAddressValidationRule;
-import org.eclipse.dataspaceconnector.gcp.dataplane.gcs.validation.ValidationRule;
 import org.eclipse.dataspaceconnector.spi.EdcException;
 import org.eclipse.dataspaceconnector.spi.monitor.Monitor;
 import org.eclipse.dataspaceconnector.spi.result.Result;
