+++ extensions/data-plane/data-plane-gcs/src/main/java/org/eclipse/dataspaceconnector/gcp/dataplane/gcs/validation/GcsSinkDataAddressValidationRule.java
@@ -14,6 +14,9 @@
 
 package org.eclipse.dataspaceconnector.gcp.dataplane.gcs.validation;
 
+import org.eclipse.dataspaceconnector.dataplane.common.validation.CompositeValidationRule;
+import org.eclipse.dataspaceconnector.dataplane.common.validation.EmptyValueValidationRule;
+import org.eclipse.dataspaceconnector.dataplane.common.validation.ValidationRule;
 import org.eclipse.dataspaceconnector.spi.result.Result;
 import org.eclipse.dataspaceconnector.spi.types.domain.DataAddress;
